+++ app/Models/Notification.php
@@ -1,18 +1,16 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class Comment extends Model
+class Notification extends Model
 {
-    protected $fillable = ['content_origin','rating','user_id','content_generated'];
+    protected $fillable = ['user_id'];
 
     protected $dates = ['deleted_at'];
 
     private $rules = array(
-        'content_origin' => 'require',
-        'content_generated' =>'require'
         // .. more rules here ..
     );
 
@@ -24,7 +22,7 @@ class Comment extends Model
         return $v->passes();
     }
 
-    public function notification()
+    public function user()
     {
         return $this->belongsTo('App\User');
     }
