+++ app/Models/Status.php
@@ -1,11 +1,14 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Status extends Model
 {
+    use SoftDeletes;
+
     protected $table = 'user_statuses';
     protected $primaryKey = 'id';
     protected $guarded = ['id'];
