+++ app/Models/VotePermission.php
@@ -1,25 +1,25 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
 
-class Vote extends Model
-
+class VotePermission extends Model
 {
+
     use SoftDeletes;
 
     public function user()
     {
         return $this->belongsTo('App\User');
     }
+    //TODO user has many VotePermission
 
-    public function topic()
+    public function vote()
     {
-        return $this->belongsTo('App\Topic');
+        return $this->belongsTo('App\Vote');
     }
 
 }
-
